:recycle: refactoring board route
